Crate wasmtime_internal_versioned_export_macros

This crate defines macros to easily define and use functions with a versioned suffix, to facilitate using multiple versions of the same crate that generate assembly.

⚠️ Warning ⚠️: this crate is an internal-only crate for the Wasmtime project and is not intended for general use. APIs are not strictly reviewed for safety and usage outside of Wasmtime may have bugs. If you’re interested in using this feel free to file an issue on the Wasmtime repository to start a discussion about doing so, but otherwise be aware that your usage of this crate is not supported.

Macros

versioned_stringify_ident

versioned_suffix

Attribute Macros

versioned_export

versioned_link